Add moderator page and update account menu
