+++ react-ui/src/components/MenuItemLink.jsx
@@ -18,8 +18,22 @@ function menuItemComponent({ state, to }) {
   ));
 }
 
+/**
+ * MenuItemLink component renders a menu item with an optional icon and text.
+ * It can be used to navigate to a specified route or trigger a click event.
+ *
+ * @param {Object} props - The properties object.
+ * @param {React.ElementType} props.Icon - The icon component to be displayed.
+ * @param {function} props.onClick - The function to be called when the menu item is clicked.
+ * @param {boolean} props.selected - Indicates if the menu item is selected.
+ * @param {Object} props.state - The state to be passed to the link component.
+ * @param {string} props.text - The text to be displayed in the menu item.
+ * @param {string} props.to - The route to navigate to when the menu item is clicked.
+ *
+ * @returns {JSX.Element} The rendered MenuItemLink component.
+ */
 function MenuItemLink({
-  Icon, onClick, state, text, to,
+  Icon, onClick, selected, state, text, to,
 }) {
   const renderLink = useMemo(
     () => (to ? menuItemComponent({ state, to }) : undefined),
@@ -27,7 +41,12 @@ function MenuItemLink({
   );
 
   return (
-    <MenuItem button component={renderLink} onClick={onClick}>
+    <MenuItem
+      button
+      component={renderLink}
+      onClick={onClick}
+      selected={selected}
+    >
       {Icon && (
         <ListItemIcon>
           <Icon />
@@ -41,6 +60,7 @@ function MenuItemLink({
 MenuItemLink.propTypes = {
   Icon: PropTypes.elementType,
   onClick: PropTypes.func,
+  selected: PropTypes.bool,
   state: PropTypes.shape({}),
   text: PropTypes.string.isRequired,
   to: types.to,
@@ -49,6 +69,7 @@ MenuItemLink.propTypes = {
 MenuItemLink.defaultProps = {
   Icon: null,
   onClick: () => {},
+  selected: false,
   state: {},
   to: null,
 };
