+++ react-ui/src/pages/index.js
@@ -1,10 +1,10 @@
-export { default as AnalyzeVotes } from './analyzeVotes/AnalyzeVotes';
 export { default as AuthorizeCallback } from './authorizeCallback/AuthorizeCallback';
 export { default as Contest } from './contest/Contest';
 export { default as Contests } from './contests/Contests';
 export { default as EntryModal } from './entry/EntryModal';
 export { default as HallOfFame } from './hallOfFame/HallOfFame';
 export { default as Home } from './home/Home';
-export { default as ReviewSubmissions } from './reviewSubmissions/ReviewSubmissions';
+export { default as AnalyzeVotes } from './mod/analyzeVotes/AnalyzeVotes';
+export { default as ReviewSubmissions } from './mod/reviewSubmissions/ReviewSubmissions';
 export { default as Settings } from './settings/Settings';
 export { default as Submission } from './submission/Submission';
