+++ react-ui/src/pages/mod/analyzeVotes/DeviationFromMean.jsx
