+++ react-ui/src/pages/mod/analyzeVotes/UserVsAverage.jsx
