+++ react-ui/src/pages/mod/analyzeVotes/functions.jsx
