+++ react-ui/src/pages/mod/analyzeVotes/markers.jsx
